[fix] Rename .jsx files to match related component name
