+++ src/components/App.jsx
@@ -3,8 +3,8 @@ import styles from '../styles/app'
 import React from 'react'
 import { translate } from '../plugins/preact-polyglot'
 
-import Sidebar from './sidebar'
-import Topbar from './topbar'
+import Sidebar from './Sidebar'
+import Topbar from './Topbar'
 
 const App = ({ t, children }) => (
   <div class={styles['fil-wrapper']}>
