+++ src/components/TopBar.jsx
@@ -2,8 +2,8 @@ import styles from '../styles/topbar'
 
 import React from 'react'
 
-import Toolbar from './toolbar'
-import Breadcrump from './breadcrumb'
+import Toolbar from './Toolbar'
+import Breadcrump from './Breadcrumb'
 
 const Topbar = () => (
   <div class={styles['fil-content-header']}>
