+++ src/main.jsx
@@ -7,8 +7,8 @@ import { render } from 'react-dom'
 import { Router, Route, Redirect, hashHistory } from 'react-router'
 import { I18n } from './plugins/preact-polyglot'
 
-import App from './components/app'
-import Table from './components/table'
+import App from './components/App'
+import Table from './components/Table'
 
 const context = window.context
 const lang = document.documentElement.getAttribute('lang') || 'en'
